R36S: type inert boot stubs for LP64 hosts

+++ port/src/n64stubs.c
@@ -23,15 +23,16 @@
 #include <PR/ultratypes.h>
 #include <PR/os.h>
 #include <tlb_manage.h>
+#include "boot.h"
 
 /* --- Segment start/end getters (normally linker-script symbols) --------- */
 /* On the PC the ROM is loaded by romdata.c; these are unused. Return 0.   */
 
-u32 get_csegmentSegmentStart(void)   { return 0; }
-u32 get_cdataSegmentRomStart(void)   { return 0; }
-u32 get_cdataSegmentRomEnd(void)     { return 0; }
-u32 get_inflateSegmentRomStart(void) { return 0; }
-u32 get_inflateSegmentRomEnd(void)   { return 0; }
+uintptr_t get_csegmentSegmentStart(void)   { return 0; }
+uintptr_t get_cdataSegmentRomStart(void)   { return 0; }
+uintptr_t get_cdataSegmentRomEnd(void)     { return 0; }
+uintptr_t get_inflateSegmentRomStart(void) { return 0; }
+uintptr_t get_inflateSegmentRomEnd(void)   { return 0; }
 
 /*
  * Linker segment boundary symbols.
@@ -59,7 +60,7 @@ u32 *_gameSegmentVaddrEnd        = NULL;
 
 /* --- Decompressor (src/inflate, not built for PC) ----------------------- */
 /* init() calls this to unpack the data segment. Unused on the PC.         */
-u32 jump_decompressfile(u32 source, u32 target, u32 buffer)
+u32 jump_decompressfile(uintptr_t source, uintptr_t target, uintptr_t buffer)
 {
     (void)source; (void)target; (void)buffer;
     return 0;
